requirements.txt
# LangChain (LATEST STABLE)
langchain>=1.0.0
langchain-core>=1.2.0
langchain-community>=0.3.0
langchain-text-splitters>=1.1.0

# Model integrations
langchain-groq>=0.2.0
langchain-nvidia-ai-endpoints>=1.0.0

# Vector DB
faiss-cpu>=1.8.0

# Utilities
streamlit>=1.32.0
python-dotenv
requests
tiktoken
gitpython
